chore: check summary for details
Added a user field to investment class
Added readable documentation to investment app serializers
Imported Package model
Created PackageView
Altered Balance fields
Made Migrations and migrate

+++ users/models.py
@@ -51,7 +51,7 @@ class CoinAddress(models.Model):
 # Sets the user's balance
 class Balance(models.Model):
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
-    balance = models.PositiveIntegerField()
+    balance = models.DecimalField(max_digits=10, decimal_places=2)
 
 class UserCryptoDetails(models.Model):
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
